requirements.txt
Calendar==1.11.4p
argparse==1.2.1
cssselect==0.8
icalendar==3.5
lxml==3.2.3
pyquery==1.2.4
python-dateutil==2.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
zope.interface==4.0.5